% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{delete_if}
\alias{delete_if}
\title{Create a SAS-style DELETE rule}
\usage{
delete_if(condition)
}
\arguments{
\item{condition}{Logical condition evaluated on the data.table.}
}
\value{
A rule object for data_step().
}
\description{
Creates a DELETE rule to remove rows from the data.table when condition is TRUE.
}
