% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{else_do}
\alias{else_do}
\title{Create a SAS-style ELSE rule}
\usage{
else_do(...)
}
\arguments{
\item{...}{Named assignments to apply when no previous IF/ELSE IF matched.}
}
\value{
A rule object for data_step().
}
\description{
Creates an ELSE rule for use inside data_step().
}
