% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{else_if_do}
\alias{else_if_do}
\title{Create a SAS-style ELSE IF rule}
\usage{
else_if_do(condition, ...)
}
\arguments{
\item{condition}{Logical condition evaluated on the data.table.}

\item{...}{Named assignments to apply when condition is TRUE.}
}
\value{
A rule object for data_step().
}
\description{
Creates an ELSE IF rule for use inside data_step().
}
