% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{if_do}
\alias{if_do}
\title{Create a SAS-style IF rule}
\usage{
if_do(condition, ...)
}
\arguments{
\item{condition}{Logical condition evaluated on the data.table.}

\item{...}{Named assignments to apply when condition is TRUE.}
}
\value{
A rule object for data_step().
}
\description{
Creates a mutually exclusive IF rule for use inside data_step().
}
